delete public css for module.css update last styles for all pages
