+++ REACT-UNIVERSOUND/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "react-pokemons-app",
+    "name": "REACT-UNIVERSOUND",
     "version": "1.0.0",
     "description": "An awesome application to handle some pokemons.",
     "dependencies": {
