--- REACT-UNIVERSOUND/public/Like_page.html
@@ -1,122 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover">
-	<title>UniverSound</title>
-	<link rel="stylesheet" href="./css/style.css">
-    <link rel="stylesheet" href="./css/like_page.css">
-    <link rel="stylesheet" href="./css/main_soundsPage.css">
-	<script src="./scripts/script.js"></script>
-	<script src="https://kit.fontawesome.com/7aba373a63.js" crossorigin="anonymous"></script>
-	<link rel="preconnect" href="https://fonts.googleapis.com">
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-<link href="https://fonts.googleapis.com/css2?family=Rubik+Bubbles&family=Young+Serif&display=swap" rel="stylesheet">
-</head>
-<body>
-    <!--page header-->
-	<header>
-		<div class="title">
-			<div>
-				<div class="logo logo_header"></div>
-			</div>
-			<nav class="nav_profil">
-				<ul class="nav_profil_ul">
-					<a href="./loader_page.html"><li><i class="fa-regular fa-user fa-lg">Account</i></li></a>
-					<a href="./signup.html"><li><i class="fa-solid fa-right-to-bracket fa-lg">Sign Up</i></li></a>
-                    <a href="./pageprincipal.html"><li><i class="fa-solid fa-right-to-bracket fa-lg">Sign Up</i></li></a>
-				</ul>
-			</nav>
-		</div>
-	</header>
-    <!--middle page-->
-	<div class="big_box">
-        <!--nav & search-->
-		<aside class="aside">
-			<nav class="nav">
-				<ul class="nav_ul">
-					<li><i class="fa-solid fa-house-crack"></i> Home</li>
-					<li><i class="fa-regular fa-bookmark"></i> Library</li>
-					<li><i class="fa-solid fa-music"></i> Playlist</li>
-				</ul>
-				<ul class="nav_ul">
-					<li><i class="fa-brands fa-searchengin"></i> Search</li>
-				</ul>
-			</nav>
-		</aside>
-        <!-- Main box-->
-		<main class="main">
-            <!-- section up-->
-            <div class="searchBox">
-                <form>
-                    <input class="styledSearch searchField" type="search" id="site-search" name="request" />
-                    <button class="styledSearch">Search</button>
-                </form>
-            </div>
-
-                <summary><h2>Liked Music</h2></summary>
-                <section class="selection">
-                 <ul class="likelist">
-                    <li class="elementlist">
-                        <img class="likepics" src="./images/djfun.jpg" alt="dj">
-                        <p>Nom du morceau<>
-                        <p>Auteur: Nom de l'auteur</p>
-                        <p>Album: Nom de l'album</p>
-                        <p>Durée: 3:30</p>
-                        <div class="boxajout croix"></div>
-                        <div class="boxajout coeur"></div>
-                      </li>
-                      <li class="elementlist">
-                        <img class="likepics" src="./images/djfun.jpg" alt="dj">
-                        <p>Nom du morceau<>
-                        <p>Auteur: Nom de l'auteur</p>
-                        <p>Album: Nom de l'album</p>
-                        <p>Durée: 3:30</p>
-                        <div class="boxajout croix"></div>
-                        <div class="boxajout coeur"></div>
-                      </li>
-                      <li class="elementlist">
-                        <img class="likepics" src="./images/djfun.jpg" alt="dj">
-                        <p>Nom du morceau<>
-                        <p>Auteur: Nom de l'auteur</p>
-                        <p>Album: Nom de l'album</p>
-                        <p>Durée: 3:30</p>
-                        <div class="boxajout croix"></div>
-                        <div class="boxajout coeur"></div>
-                      </li>
-                 </ul>
-                </section>
-
-            <!-- section down-->
-
-		</main>
-	</div>
-    <!--page footer-->
-	<footer class="footer">
-		<nav class="nav_footer">
-			<hr>
-			<div class="logo logo_footer"></div>
-			<hr>
-			<ul class="RGPD box_footer">
-				<li>General Condition</li>
-				<li>Gestion des Cookies</li>
-				<li>Mentions légales</li>
-			</ul>
-			<hr>
-			<ul class="user box_footer">
-				<li>Profil</li>
-				<li>Last Sound</li>
-				<li>Playlist</li>
-			</ul>
-			<hr>
-			<ul class="contact box_footer">
-				<li>About Us</li>
-				<li>Contact</li>
-				<li>Inovation</li>
-			</ul>
-			<hr>
-		</nav>
-		<p>create By Team Sound for Holberton</p>
-	</footer>
-</body>
-</html>
