--- REACT-UNIVERSOUND/public/css/login.css
@@ -1,63 +1,0 @@
-main {
-    display: flex;
-    flex-direction: column;
-    background-color: black;
-    width: 50vw;
-    padding: 3rem 15% 3rem 15%;
-    border-radius: 10px;
-
-}
-h2 {
-    text-align: center;
-    margin-bottom: 20px;
-}
-.field {
-    display: flex;
-    flex-direction: column;
-}
-.g-signin2 {
-    display: inline-block;
-    background-color: #fff;
-    padding: 10px;
-    color: #fff;
-    border-radius: 5px;
-    cursor: pointer;
-    margin: 10px;
-}
-form {
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-label {
-    margin-bottom: 8px;
-    display: flex;
-}
-
-input, select, textarea {
-    width: 100%;
-    padding: 8px;
-    margin-bottom: 16px;
-}
-#remember {
-    width: 10%;
-}
-.Remem {
-    margin-left: 3px;
-}
-
-button {
-    background-color: var(--creamy);
-    color: #000000;
-    padding: 10px 15px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: bold;
-}
-
-button:hover {
-    background-color: #fff;
-}
-
