--- REACT-UNIVERSOUND/public/css/main_soundsPage.css
@@ -1,167 +1,0 @@
-body {
-	background: linear-gradient(
-
-	rgba(0, 0, 0, 0.877), 
-
-	rgba(0, 0, 0, 0.63)
-
-  ),url(../images/waveC.jpg)no-repeat;background-size:cover;}
-
-.main {
-	display: flex;
-	flex-direction: column;
-    width: 100%;
-    padding: 1rem;
-	font-family: 'Young Serif', 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-	flex-wrap: wrap;
-}
-/*hidden our selection*/
-summary {
-	display: block;
-	margin-bottom: 20px;
-}
-summary:hover {
-	color: rgba(138, 43, 226, 0.5);
-
-}
-/* selection screen*/
-.selection {
-	border: 2px solid ;
-	border-radius: 10px;
-	padding: 20px;
-	margin-bottom: 30px;
-	border: solid 2px var(--neon-blue);
-	box-shadow: 2px 2px 10px var(--neon-blue-shadow);
-	flex-wrap: wrap;
-	max-width: 1000px
-}
-.selection h2 {
-	margin-left: 25px;
-	
-}
-/*.nav_sounds_selection {
-	display: grid;
-	grid-template-columns: repeat(5, 1fr);
-	grid-template-rows: repeat(2, 1fr);
-	gap: 10px;
-}*/
-.nav_sounds_selection {
-	display: flex;
-	justify-content: space-around;
-	flex-wrap: wrap;
-}
-
-/*personalisation sounds*/
-.yourStyle {
-	display: flex;
-	justify-content: space-around;
-	width: 100%;
-	flex-wrap: wrap;
-}
-.box_yourstyle {
-	padding: 20px;
-	margin-bottom: 40px;
-	border: 2px solid;
-	border-radius: 10px;
-	border: solid 2px var(--neon-blue);
-	box-shadow: 2px 2px 10px var(--neon-blue-shadow);
-}
-.title_box {
-	text-align: center;
-}
-/*jacket box*/
-.SoundBox{
-	display: flex;
-	margin: 1rem;
-	padding: 20px;
-	width: 140px;
-	height: 200px;
-	justify-content: center;
-	border-radius: 10px;
-	background: linear-gradient(to bottom, #292929, #1a1a1a,  rgb(17, 17, 17));
-}
-.SoundBox:hover {
-	background: linear-gradient(to bottom, #303030, #202020,  #292929);
-	box-shadow: 2px 2px 10px rgb(58, 57, 57);
-	color: azure;
-}
-.SoundBox:hover .player {
-	display: block;
-	transition: opacity 1s ease-in-out;;
-}
-.jacket_box {
-	width: 100px;
-	position: relative;
-	border-radius: 10px;
-}
-.jacket {
-	height: 100px;
-	width: 99px;
-	border-radius: 10px;
-}
-.player {
-	display: none;
-	height: 20px;
-	width: 20px;
-	position: absolute;
-	right: 0;
-	top: 0;
-	transition: transform 0.5s ease-in-out;
-}
-.player:hover {
-	transform: scale(1.1);	
-}
-.jacket_box h4 {
-	font-size: 13px;
-	margin-left: 4px;
-}
-.jacket_box p {
-	font-size: 10px;
-	margin-left: 4px;
-}
-
-/*Search bar*/
-.searchBox {
-	display: flex;
-	justify-content: center;
- 	background-color: rgba(77, 26, 216, 0.13);
-	border: solid 2px var(--neon-blue);
-	box-shadow: 2px 2px 10px var(--neon-blue-shadow);
-	border-radius: 10px;
-	width: 100%;
-	margin-bottom: 20px;
-}
-.styledSearch {
-	border: 0;
-	line-height: 2.5;
-	padding: 0 20px;
-	font-size: 1rem;
-	text-align: center;
-	color: #fff;
-	text-shadow: 1px 1px 1px #000;
-	border-radius: 10px;
-	background-color: rgb(22, 66, 133);
-	background-image: linear-gradient(
-	  to top left,
-	  rgb(0 0 0 / 20%),
-	  rgb(0 0 0 / 20%) 30%,
-	  rgb(0 0 0 / 0%)
-	);
-	box-shadow:
-	  inset 2px 2px 3px rgb(255 255 255 / 60%),
-	  inset -2px -2px 3px rgb(0 0 0 / 60%);
-  }
-  
-  .styledSearch:hover {
-	background-color: rgba(34, 17, 105, 0.76);
-  }
-  
-  .styledSearch:active {
-	box-shadow:
-	  inset -2px -2px 3px rgb(255 255 255 / 60%),
-	  inset 2px 2px 3px rgb(0 0 0 / 60%);
-  }
-  .searchField {
-	width: 20vw;
-  }
-
