--- REACT-UNIVERSOUND/public/css/style.css
@@ -1,89 +1,0 @@
-:root {
-    --neon-blue: #00bfff9f;
-    --neon-purple: #892be293;
-	--backg-dark: #000000;
-    --neon-blue-shadow: rgba(0, 102, 255, 0.5);
-    --neon-purple-shadow: rgba(138, 43, 226, 0.5);
-    --creamy: rgb(218, 197, 171);
-}
-*, ::before, ::after {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-}
-
-/* General page custom */
-html {
-    margin: 0;
-    padding: 0;
-}
-
-body {
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    margin: 0;
-    padding: 0;
-    background: linear-gradient(to bottom, black,  #1a1a1a, #292929, #1a1a1a, black);
-    
-}
-ul {
-    list-style: none;
-}
-
-/* Import images */
-.logo {
-    background-image: url('../images/logov2Trans.png');
-	background-repeat: no-repeat;
-	background-position: center;
-    height: 101px;
-    width: 224px;
-	
-}
-.logo_header {
-	box-shadow:  inset -5px -5px 15px rgba(112, 112, 228, 0.5), 10px 1px 10px rgba(112, 112, 228, 0.5);
-}
-
-
-/* Header */
-.title {
-    display: flex;
-    justify-content: space-between;
-    border-bottom: 2px solid var(--neon-purple);
-    padding: 1rem;
-	box-shadow: 0 0 10px var(--neon-purple-shadow);
-}
-
-.nav_profil {
-    display: flex;
-    align-items: center;
-    width: 25%;
-}
-
-.nav_profil_ul {
-    display: flex;
-    padding: 0;
-    width: 100%;
-    justify-content: space-evenly;
-}
-.nav_profil_ul li {
-    color: #8b98df;
-}
-.nav_profil_ul li:hover{
-    color: var(--neon-blue);
-}
-
-/* Big box */
-.big_box {
-	color:azure;
-    display: flex;
-    justify-content: center;
-    height: auto;
-    background-image: url('');
-    background-repeat: no-repeat;
-    background-size: cover;
-	margin-top: 30px;
-    margin-bottom: 30px;
-}
-
-
