--- REACT-UNIVERSOUND/public/loading.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>UniverSound</title>
-    <link rel="stylesheet" href="./styles/loader.css">
-    <script src="./scripts/loader.js"></script>
-</head>
-<body>
-    <!-- audio hearthbeat-->
-    <audio id="heartbeatAudio" autoplay controls>
-        <source  src="./sounds/bup.mp3" type="audio/mp3">
-    </audio>
-    <main>
-        <!--logo-->
-        <div class="logo-container">
-            <img src="./images/logov2Trans.png" alt="Votre logo">
-        </div>
-    </main>
-</body>
-</html>
