+++ REACT-UNIVERSOUND/src/app.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import LandingPage from './pages/landingPage/landingPage';
 import { BrowserRouter as Router, Link, Switch, Route } from 'react-router-dom'
-
 import SoundPage from './pages/soundPage/soundPage';
 import SignUp from './pages/userLogPage/signuppage/signup';
 import LoginPage from './pages/userLogPage/loginpage/login';
