+++ REACT-UNIVERSOUND/src/components/asideBox/asideNav/asideMenu.module.css
@@ -16,14 +16,19 @@
     padding: 0;
     margin: 0;
     height: 200px;
-	
+
 }
 .nav_ul li {
-	margin: 30px;
+    margin: 30px;
 }
-.nav_ul li:hover {
+.nav_link {
+    text-decoration: none;
+    color: white;
+}
+.nav_link:hover {
     border: 1px solid #fff;
     border-radius: 10px;
     background-color: aliceblue;
-	color: rgb(0, 0, 0);
+	color: black;
+    padding: 0 20px;
 }
