+++ REACT-UNIVERSOUND/src/components/footer/footer.module.css
@@ -33,7 +33,7 @@
 }
 .footer ul {
 	text-align: center;
-    
+
 }
 
 .logo {
@@ -44,7 +44,10 @@
     width: 224px;
 }
 
+.box_footer a {
+
+    color: white;
+}
 a {
     text-decoration: none;
-    color: white;
 }
