+++ REACT-UNIVERSOUND/src/components/main/songPageBox/coverBox/songCover.module.css
@@ -1,0 +1,66 @@
+/*jacket box*/
+
+.SoundBox{
+	display: flex;
+	margin: 1rem;
+	padding: 20px;
+	width: 140px;
+	height: 200px;
+	justify-content: center;
+	border-radius: 10px;
+	background: linear-gradient(to bottom, #292929, #1a1a1a,  rgb(17, 17, 17));
+}
+.SoundBox:hover {
+	background: linear-gradient(to bottom, #303030, #202020,  #292929);
+	box-shadow: 2px 2px 10px rgb(58, 57, 57);
+	color: azure;
+}
+.SoundBox:hover .player {
+	display: block;
+	transition: opacity 1s ease-in-out;;
+}
+.jacket_box {
+	width: 100px;
+	position: relative;
+	border-radius: 10px;
+}
+.jacket {
+	height: 100px;
+	width: 99px;
+	border-radius: 10px;
+}
+.player {
+	display: none;
+	height: 20px;
+	width: 20px;
+	position: absolute;
+	right: 0;
+	top: 0;
+	transition: transform 0.5s ease-in-out;
+}
+.player:hover {
+	transform: scale(1.1);
+}
+.jacket_box h4 {
+	font-size: 13px;
+	margin-left: 4px;
+}
+.jacket_box p {
+	font-size: 10px;
+	margin-left: 4px;
+}
+.yourStyle {
+	display: flex;
+	justify-content: space-around;
+	width: 100%;
+	flex-wrap: wrap;
+}
+.box_yourstyle {
+	padding: 20px;
+	margin-bottom: 40px;
+	border: 2px solid;
+	border-radius: 10px;
+	border: solid 2px var(--neon-blue);
+	box-shadow: 2px 2px 10px var(--neon-blue-shadow);
+}
+
