+++ REACT-UNIVERSOUND/src/components/mainV3/songPage/song-list-historic.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent, useState, useEffect } from "react";
 import SONGS from "../../../models/mock-song";
-import SongCover from "../../../components/main/coverBox/song-cover";
+import SongCover from "../../main/songPageBox/coverBox/song-cover";
 import Song from "../../../models/song";
 
 const SongListHistoric: FunctionComponent = () => {
@@ -20,7 +20,7 @@ const SongListHistoric: FunctionComponent = () => {
                 <SongCover key={song.id} song={song}/>
             ))}
        </nav>
-    </section>                                    
+    </section>
   );
 }
