+++ REACT-UNIVERSOUND/src/components/mainV3/songPage/song-type-list.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent, useState, useEffect } from "react";
 import SONGS from "../../../models/mock-song";
-import SongCover from "../../../components/main/coverBox/song-cover";
+import SongCover from "../../main/songPageBox/coverBox/song-cover";
 import Song from "../../../models/song";
 
-const SongRhithmList: FunctionComponent = () => {
+const SongTypeList: FunctionComponent = () => {
   const [songs, setSongs] = useState<Song[]>([]);
 
   useEffect(() => {
@@ -14,14 +14,14 @@ const SongRhithmList: FunctionComponent = () => {
 
   return (
      <div className="box_yourstyle">
-         <header className="title_box"><h3>Same Rhythm</h3></header>
+         <header className="title_box"><h3>Same Style</h3></header>
          <nav className="nav_sounds_user nav_sounds_lastlistening">
           {lastSongs.map(song => (
                 <SongCover key={song.id} song={song}/>
             ))}
          </nav>
-    </div>                
+    </div>
   );
 }
 
-export default SongRhithmList;
+export default SongTypeList;
